Initial "debian" directory
Let's start with whatever dh_make suggested to put into "debian"
subdirectory. We branch this commit off the 0.8 tag. The idea is
to keep the history something like this:

       tag: 0.8          tag: 0.8.1
       '                 '
    ---o---o---o-- ... --o---o        master
        \                 \
         o---o---o-- ... --o--o--o    debian/sid
                 '               '
                 tag: 0.8.0-1    tag: 0.8.1-1

So Debian packaging is a long-lived branch, parallel to the main
"master" branch with upstream code of Chibi-Scheme. The main package
development is happening in "debian/sid" branch, with branches for
other Debian versions, maybe.