Helper script for original tarballs
Chibi-Scheme is not Debian-native package, so it requires an original
tarball present in the parent directory. The source archive should have
a particularly versioned name.

Let's add a helper script to create such tarball from whatever the
current version of Chibi is. Running "debian/make-orig-tarball" will
produce a suitable tarball. We take care to make it reproducible.