Correct installation paths
Chibi-Scheme does not use Autotools -- and that's good! -- but debhelper
makes it easy only for Autotols. dh_auto_configure passess all required
flags with Debian-specific paths to ./configure... which does nothing.
It is expected to... um... configure the Makefile to use these paths.

We'd like to piggyback on dh_auto_configure to figure out the correct
paths for us, so we resort to a vile hack: replace the ./configure
script with a helper that writes out all the interesting variables
to a temporary file, and then those variables are passed to "make"
when building and installing, thus 'configuring' the installation.

That's that the first patch does. Note that the variables used by Chibi
are different from standard GNU variables. The standard ones are all
lowercase (and that's what dh_auto_configure gives us). We convert them
to uppercase as expected by Chibi.

Also note the second patch which makes sure that overrides set by
dh_auto_configure are actually applied by Chibi's Makefile. It installs
libraries and headers via non-GNU-standard variables that share common
prefix.

In the end, all of this allows to use dh_auto_* targets with minimal
overrides. One last thing is that the static library is not included
into "all" target by default so we have to add it to the build manually.
It is installed only if it is present in the build tree.