Improve testing step
First of all, fix a typo in dh_auto_test override. After that tell it to
use the "test-all" target which runs a bit more tests. There is also a
"test-dist" target, but it's currently broken due to issues with static
linkage. Chibi-Scheme has to be specially compiled to support static
linkage required by memory tests. (Thus I also suspect that the static
library we're currently building is actually unusable.)