Clean working directory for original tarball
The packing script does not make sure that working directory is clean.
This can cause various Debian crap to be packed into the 'pristine'
tarball. For example, the ".pc" directory created by Quilt when patching
the source code.

I've been burnt by this more than once and want it to stop. Let's clean
the directory and make sure we use the state as recorded in Git history.