Fix *.meta file generation
Debian packages are usually built with parallel make (with -j<num-cpus>
flag). This requires make prerequisites to be completely accurate. Chibi
gets it mostly right, but this particular case was broken. "make all"
works in a single-threaded mode because of the target ordering, but with
parallel builds we have to explicitly specify that we need Chibi binary
to generate the *.meta files.