Properly build static library
Debian policy prescribes that the static libraries in -dev packages are
intended for statically-linked applications. Therefore they must be
compiled without -fPIC and without dlopen() usage.

Thankfully, Chibi makes it relatively easy to achieve. First we compile
the static library with the flags we need, then we clean the build (this
removes object files, but not the library), and then we do a regular
build to have a complete set of files.