Accurate binary dependencies
First of all, fixup "depends" fields. We probably don't have to specify
the minimum version of libchibi-scheme0 and libchibi-scheme-ffi, but
since we don't track symbol changes yet, it's a good precaution (and
it signals that this version may be bumped when necessary). Also note
that libchibi-scheme-dev requires libchibi-scheme0 of exactly the same
version to ensure that we link to the proper library and use the correct
headers.

Add chibi-scheme-common as a recommended package for chibi-scheme. While
it's possible to run chibi-scheme without a standard library (with -Q),
that's probably not what you want. However, this is not hard dependency,
hence it's "recommends", not "depends".

Finally, throw in the docs as suggested package for chibi-scheme binary
and the developement libraries. Human users should find them useful.