Initial improved man pages
The ones that are bundled with the upstream are not bad, but we'd
like the man pages to be a better quality. I have updated existing
ones, tweaked them to be more like the other in the system. Also,
snow-chibi binary does not have a man page at all. Now it has.

The man pages are installed with dh_installman. They overwrite the
ones installed by the upsteam. We keep the upstream installation
for the sake of simplicity.