lintian overrides for extension libraries
lintian produces a couple of warnings for extension libraries
that Chibi bundles. After careful examination I have found those
to be false positives so I'm overriding them.

profile.so is really linked only against libchibi-scheme.so.0
and does not use any symbols from libc:

$ nm -D lib/chibi/optimize/profile.so
                 w __cxa_finalize
                 w __gmon_start__
                 w _ITM_deregisterTMCloneTable
                 w _ITM_registerTMCloneTable
                 U sexp_define_foreign_aux
0000000000001160 T sexp_init_library
                 U sexp_type_exception

emscripten.so does not even use that. It exports a single function
that does call anything, it just does some trivial arithmetic and
memory access.

That's generated code for you and extension libraries for scripting
languages. These tend to be fairly minimal as they are not general-
purpose shared libraries.