Better Bash completions
Chibi ships Bash completions in contrib directory, but they don't cover
all binaries installed by Chibi. Let's improve this and provide them for
every tool that we ship.

Writing completions turned out to be quite fun -- AND they tremendously
improve discoverability and usability of command-line tools. Win-Win.

The implementation is mostly complete for 0.8, but there are some warts:
Snow is awfully slow so are completions of package names. In some cases
we even don't provide any completions at all because the API is lacking.

I have no idea whether this works with zsh. It can be ported, I guess,
like git does it. However, I have tested it on macOS with their standard
utilities and completions seem to work there too, so that's nice.