Automatic packaging tests
Drop in some tests for "autopkgtest" that verify that the package seems
to be usable. These are just low-effort smoke tests but they do the job.

I have to say that autopkgtest is cool. It was easy to set up a VM and
develop all these little testies. I've done all of this in like 3 hours.