Update Debian changelog
Since the previous version has not been released, moved the ITP closure
to the new changelog entry, update it for the new version.

Also update the control files to declare dependency on newer shared
library packages. Instead of hardcoding the versings, risking to forget
to update them, just use the current package version.