Don't warn about missing config file
This reverts commit c3189ebc9d70ebfb44d9278ae25fb61cf2fc4ef6
("warn when config fails to load", 2020-01-06).

The commit has introduced a warning message when configuration file
cannot be read. However, if the configuration file is missing
altogether, the default value should be returned without any
additional annoying output.

In particular "snow-chibi" uses a configuration file in the home
directory (~/.snow/config.scm) which is not created by default.
If it is missing, every Snow command outputs a warning.

It's not very convenient to inspect the exception to see if it is
caused by a missing path (ENOENT) or something else, so for the time
being let's just revert this change.