Fix stack traces in prebuilt Chibi images
One thing that I've noticed after enabling image support for Snow-Chibi
is that the stack traces contain Debian DESTDIR path, not the actual
system path. This is not nice.

Unfortunately, the issue itself stems from the fact that Chibi bytecode
remembers the original paths of the files used to generate images. There
is no easy way to hack that from inside Chibi.

Therefore, we take roundabout approach: use chroot to get Chibi to
believe that it's already installed into the user system when generating
images. That way the bytecode will contain correct paths.

There are some tricky details to this approach which you can read in the
attached patch comment, but other than that it works pretty well.