Man page alias for "snow-chibi.scm"
Recent reactivation of image installation makes "make install" to
install "snow-chibi.scm" to /usr/bin, which is a standalone executable
script on its own. Lintian complains about a binary without man page
so let's add one.

We need "snow-chibi.scm" to have the same man page as "snow-chibi".
There is not-so-well-documented feature of "dh_installman" called
".so links" or ".so aliases" which allows to conveniently install
a symlink to the manpage.