Revert "Fix stack traces in prebuilt Chibi images"
This reverts commit df97bb1d8930db8137ecdbcd0582db93c856cf26.

It seems that "fakechroot" has weird interaction with *actual* chroots
used to build packages in clean environment. While it seems to be
working on my machine, trying to build the package with "pdebuild"
results in weird errors about "libchibi-scheme.so.0" being not found.
When inspecting dependencies with "ldd", I see this:

    fakechroot chroot /build/chibi-scheme-0.9.1/debian/tmp /usr/bin/ldd /usr/bin/chibi-scheme
            linux-vdso.so.1 (0x0000000000000000)
            libfakechroot.so => /usr/lib/x86_64-linux-gnu/fakechroot/libfakechroot.so (0x0000000000000000)
            libchibi-scheme.so.0 => not found
            libc.so.6 => /lib/x86_64-linux-gnu/libc.so.6 (0x0000000000000000)
            /lib64/ld-linux-x86-64.so.2 (0x0000000000000000)
    fakechroot chroot /build/chibi-scheme-0.9.1/debian/tmp /usr/bin/chibi-scheme -d /usr/share/chibi/chibi.img
    /usr/bin/chibi-scheme: error while loading shared libraries: libchibi-scheme.so.0: cannot open shared object file: No such file or directory
    make[2]: [Makefile:401: install] Error 127 (ignored)

The loader seems to ignore the fake chroot and sits in the real one.
Well, I'd say that the approach with fakechroot is not going to fly
and we need a different way to fix up the stack trace paths. (Never
liked the idea of this hack, actually. Since it's *a hack* after all.)