Add Multi-Arch annotations
Multiarch hinter hints are mostly correct.

    chibi-scheme          can't be co-installed, binaries       (*)
    chibi-scheme-common   "foreign", arch-independent           (*)
    chibi-scheme-doc      "foreign", arch-independent
    chibi-scheme-images   can't be co-installed, bytecode       (*)
    libchibi-scheme-dev   "same", can be co-installed per-arch
    libchibi-scheme-ffi   "same", can be co-installed per-arch
    libchibi-scheme0      "same", can be co-installed per-arch

chibi-scheme contains "chibi-scheme" binary, it can't be co-installed.

chibi-scheme-common by itself contains only arch-independent Scheme
code. However it depends on libchibi-scheme-ffi with native libraries.

chibi-scheme-images generates bytecode in postinstall scripts, this
bytecode is architecture-dependent and cannot be co-installed. (Chibi's
bytecode per se is 'portable' but the feature flags might be different
on platforms, so bytecode generated for x86_64 might not be loadable
on arm, for example.)