Don't hardcode dependency versions for chibi-scheme-common
chibi-scheme-common depends on libchibi-scheme-ffi to pull in required
corresponding native libraries without which Scheme libraries won't load.

We need the FFI package to be built at least from the same version but
it might be something newer. Currently there was no precendent for
libraries being removed and the ABI is stable, so Scheme code for 0.9.1
should work with 0.10, for example.

However, we don't really want to hardcode this version exactly. Use the
version of the source package (0.9.1) to not depend on the exact binary
package version (e.g., 0.9.1-3).