Backport upstream fixes to the test suite
Debian's test runs are failing on some architectures that aren't x86_64.
Backport a bunch of patches from upstream to resolve those issues.

All of these have been submitted to upstream (by me), some of them made
it into already available 0.10, some are probably going to float on the
development branch for a while.

For now, it's easier to backport these onto packaged 0.9.1 to get the
builds all nice and clean. Patch #16 can be dropped when 0.10 is
packaged, but we'll likely need to carry around for a while.